requirements/base.txt
# Base
pytz==2018.7
python-slugify==1.2.6
Pillow==5.3.0
psycopg2==2.7.4 --no-binary psycopg2

# Django
django==2.0.9

# Django REST Framework
djangorestframework==3.9.2

# Json Web Tokens
pyjwt==1.7.1
djangorestframework_simplejwt==4.1.1

# Environment
django-environ==0.4.5

# Passwords security
argon2-cffi==18.3.0

# Static files
whitenoise==4.1.2

